Compile debugged with gcc-4.7.  make check runs successfully.

+++ fiber_bundles_test/examples/fields3d.t.cc
@@ -120,7 +120,7 @@ Dof_tuple make_coord_field_dofs(int edge_ct_x, int edge_ct_y)
     }
   }
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
 
@@ -182,7 +182,7 @@ Dof_tuple make_scalar_field_dofs(int edge_ct_x, int edge_ct_y, int step)
     }
   }
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
 
@@ -359,7 +359,7 @@ make_t2_e3_field_dofs(int xedge_ct_x, int xedge_ct_y)
   }
 
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
 
@@ -476,7 +476,7 @@ make_st2_e3_field_dofs(int xedge_ct_x, int xedge_ct_y)
   }
 
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
 
@@ -578,7 +578,7 @@ make_at2_e3_field_dofs(int xedge_ct_x, int xedge_ct_y)
   }
 
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
 
@@ -731,7 +731,7 @@ make_t3_e3_field_dofs(int xedge_ct_x, int xedge_ct_y)
   }
 
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
