+++ fiber_bundles_test/section_spaces/section.t.cc
@@ -89,7 +89,7 @@ make_scalar_field_dofs(int edge_ct_x, int edge_ct_y)
     }
   }
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
 
@@ -140,7 +140,7 @@ make_coord_field_dofs(int edge_ct_x, int edge_ct_y)
     }
   }
 
-  Dof_tuple result = {dofs, dofs_ub*sizeof(double)};
+  Dof_tuple result = {dofs, dofs_ub*static_cast<int>(sizeof(double))};
 
   // Postconditions:
